GunSmoke: Char colours OK. Background visible but mapping wrong. No objects

+++ bin/jtlinks.sh
@@ -2,14 +2,14 @@
 
 # Creates a reversed folder structure
 
-cd $JTGNG_ROOT
-
-for i in hdl mist mister; do
+for i in hdl mist mister ver; do
     mkdir -p $JTGNG_ROOT/$i
     cd $JTGNG_ROOT/$i
     for j in 1943 1942 commando gunsmoke; do
-        ln -f -s $JTGNG_ROOT/$j/$i $j
+        if [ -e $JTGNG_ROOT/$j/$i ]; then
+            ln -f -s $JTGNG_ROOT/$j/$i $j
+        fi
     done
+    echo $JTGNG_ROOT/$i
 done
 
-echo "Symbolic links created"
